Organized media menu code
