+++ modules/menus/media/components/controls/playpause/helpers.ts
@@ -1,0 +1,4 @@
+import { PlaybackStatus } from 'lib/types/mpris';
+
+export const isValidPlaybackStatus = (status: string): status is PlaybackStatus =>
+    ['playing', 'paused', 'stopped'].includes(status);
